fix elu output type

+++ onnx/reference/ops/op_elu.py
@@ -9,4 +9,4 @@ from ._op import OpRunUnaryNum
 class Elu(OpRunUnaryNum):
     def _run(self, x, alpha=None):  # type: ignore
         alpha = alpha or self.alpha  # type: ignore
-        return (np.where(x > 0, x, alpha * (np.exp(x) - 1)),)
+        return (np.where(x > 0, x, alpha * (np.exp(x) - 1)).astype(x.dtype),)
